KS 11_3. Задача 3 к Уроку 11
    Внес изменения:
    Функции внес в тело класса Room и вызывал у обьектов.
    Удалил лишнее.
