+++ src/main/kotlin/lesson11/task3/Participant.kt
@@ -1,0 +1,7 @@
+package lesson11.task3
+
+class Participant(
+    var speakerStatus: String,
+    val nikName: String,
+    val avatar: String,
+)
