+++ src/main/kotlin/lesson11/task3/lesson11_task3.kt
@@ -1,0 +1,44 @@
+package lesson11.task3
+
+fun main() {
+    val room =
+        Room(
+            cover = "red",
+            nameRoom = "Комната для любителей бега",
+            listParticipants = mutableListOf(),
+        )
+
+    val participant =
+        Participant(
+            speakerStatus = "mute",
+            nikName = "Denis",
+            avatar = "парень",
+        )
+
+    room.addingParticipant(participant, room.listParticipants, room)
+    room.clickAvatar(participant)
+    room.updateStatus(participant)
+    room.clickAvatar(participant)
+}
+
+// Прокачиваем абстрактное мышление. Задача на проектирование сущностей для
+// упрощенной версии приложения социальной сети, в которой общаются только голосом.
+// Требуется описать сущности и имитировать действия методами.
+//
+// Функционал одного из модулей будет такой. Есть лента, в которой отображаются
+// “комнаты” для общения по интересам. В карточке “комнаты” отображается:
+//
+// - обложка;
+// - название;
+// - список участников — отображаются в виде аватарок. При долгом нажатии на
+// аватар подсветится его никнейм. Рядом с аватаркой отображается текстовый бейдж
+// с одним из возможных статусов говорящего: “разговаривает”, “микрофон выключен”,
+// “пользователь заглушен”.
+//
+// Класс должен содержать методы:
+//
+// – добавления участника (принимает объект участника и сохраняет в свойство комнаты);
+// – обновления статуса (принимает имя пользователя и новый статус);
+// – вывода информации о пользователе при долгом нажатии на аватар.
+//
+// После проектирования создай объект комнаты с произвольными данными.
